backend/requirements.txt
autopep8
decorator
Django==4.2.3
django-environ
django-extensions
django-redis
idna
redis
requests

celery
django-celery-beat
flower

djangorestframework
djangorestframework-simplejwt
pushpad
psycopg2
humanize
django-model-utils
django-countries
django-money
numpy
coreapi
# django-language-fields
django-admin-volt
django-phonenumber-field[phonenumbers]
# Werkzeug
python-telegram-bot==13.7
django-background-tasks-updated


https://github.com/django-mptt/django-mptt/archive/refs/heads/main.zip
https://github.com/django-polymorphic/django-polymorphic/archive/refs/heads/master.zip
django-rest-polymorphic
https://github.com/django-polymorphic/django-polymorphic-tree/archive/refs/heads/master.zip
https://github.com/viewflow/django-fsm/archive/refs/heads/master.zip
https://github.com/jazzband/django-fsm-log/archive/refs/heads/master.zip
https://github.com/gadventures/django-fsm-admin/archive/refs/heads/master.zip
djangoql
pydotplus
pyparsing
pydot
graphviz
django-custom-query
django-nested-admin
django-viewflow
django-material
django-jsonform
django-better-json-widget
djsingleton


django-mathfilters
django-admin-volt

pyyaml
ua-parser
user-agents

beautifulsoup4
django-ipware



pygraphviz


django-language-field





Pygments

drf-spectacular==0.27.1


sentry-sdk


numexpr


dnspython

docker


langchain==0.1.5
openai==1.11.1
langchain-community
langchainhub


django-cors-headers


Faker

phone-gen

webauthn
django-filter